fixed gui issues
